extract stats and notes
