+++ services/__tests__/stats.test.ts
@@ -261,11 +261,6 @@ describe('Stats management', () => {
   });
 
   describe('getStatsForDate', () => {
-    it('should return null when no stats exist for date', async () => {
-      const stats = await getStatsForDate('2024-03-15');
-      expect(stats).toBeNull();
-    });
-
     it('should return stats for specific date', async () => {
       await updateStats(Rating.Good, false);
       await updateStats(Rating.Hard, true);
